micro: split workspace browser events
